remove references to sf_hostHasService
we can no longer support the HostPropertySyncJob that fills out this field and thus need to move away from it.  showing infrastructure stats against a specific plugin now does an all-host join against a known emitted metric.  this can be less accurate at host counts greater than job fetch limits, and is probably a lot slower.  all other usages are unncessary and have been removed.  note that "hostHasService" is not a problem as it is merely a dimension emitted by a plugin, much like a second tier "plugin" value

https://signalfuse.atlassian.net/browse/SQUID-781